Integrate robot testing directly into desktop.rs framework
Robot support is now integrated into THE actual app, not a separate
emulation. This eliminates code duplication and ensures tests run
against the real production code path.

Key changes:
- Added RobotCommand enum and RobotAppHandle to desktop.rs (behind robot feature)
- Implemented run_with_robot() that uses EventLoop<RobotCommand>
- Robot commands handled in same event loop as normal window events
- Same WGPU setup, same rendering, same input handling - just with programmatic control added
- Deleted desktop_robot.rs (was duplicate event loop code)
- Updated robot_app.rs to use integrated desktop module

API:
- Normal: desktop::run(settings, content) -> !
- Robot: desktop::run_with_robot(settings, content) -> RobotAppHandle

The robot mode runs THE SAME app with all production code paths,
just adds robot command handling via user events.
